+++ regression-tests/long_double/add_2.c
@@ -1,6 +1,6 @@
 int main() {
-  double a = 3.0f;
-  double b = 0.5f;
-  assert(a+b == 3.5f);
+  long double a = 3.0l;
+  long double b = 0.5l;
+  assert(a+b == 3.5l);
   return 0;
 }
