+++ regression-tests/long_double/add_3.c
@@ -1,5 +1,5 @@
 int main() {
-  double a = 2.0f;
-  double b = a + 0.5f;
-  assert(a + 0.4f == b);
+  long double a = 2.0l;
+  long double b = a + 0.5l;
+  assert(a + 0.4l == b);
 }
